[ArticleCategories] move to directory and separate item from ArticleCategories to make room for creation & upvote/downvote
