+++ components/ArticleCategories/ArticleCategory.js
@@ -1,24 +1,19 @@
 import { useState } from 'react';
 import gql from 'graphql-tag';
-import { t } from 'ttag';
-import { useQuery, useMutation } from '@apollo/react-hooks';
+import { useMutation } from '@apollo/react-hooks';
 import Chip from '@material-ui/core/Chip';
 import Menu from '@material-ui/core/Menu';
 import MenuItem from '@material-ui/core/MenuItem';
-import Button from '@material-ui/core/Button';
 import { makeStyles } from '@material-ui/core/styles';
-import AddIcon from '@material-ui/icons/Add';
 
 import { dataIdFromObject } from 'lib/apollo';
 
 const useStyles = makeStyles(theme => ({
-  button: { marginTop: theme.spacing(1) },
-  buttonIcon: { marginRight: theme.spacing(1) },
   category: { marginRight: theme.spacing(1), marginTop: theme.spacing(1) },
 }));
 
-const ArticleCategoriesData = gql`
-  fragment ArticleCategoriesData on ArticleCategory {
+const ArticleCategoryData = gql`
+  fragment ArticleCategoryData on ArticleCategory {
     # articleId and categoryId are required to identify ArticleCategory instances
     articleId
     categoryId
@@ -33,8 +28,8 @@ const ArticleCategoriesData = gql`
   }
 `;
 
-const ArticleCategoriesDataForUser = gql`
-  fragment ArticleCategoriesDataForUser on ArticleCategory {
+const ArticleCategoryDataForUser = gql`
+  fragment ArticleCategoryDataForUser on ArticleCategory {
     # articleId and categoryId are required to identify ArticleCategory instances
     articleId
     categoryId
@@ -46,24 +41,10 @@ const ArticleCategoriesDataForUser = gql`
 const ArticleWithCategories = gql`
   fragment ArticleWithCategories on Article {
     articleCategories {
-      ...ArticleCategoriesData
+      ...ArticleCategoryData
     }
   }
-  ${ArticleCategoriesData}
-`;
-
-const CATEGORY_LIST_QUERY = gql`
-  query ListCategoryForSelect {
-    ListCategories(first: 50) {
-      edges {
-        node {
-          id
-          title
-          description
-        }
-      }
-    }
-  }
+  ${ArticleCategoryData}
 `;
 
 const DELETE_CATEGORY = gql`
@@ -83,7 +64,7 @@ const DELETE_CATEGORY = gql`
   }
 `;
 
-function Category({ articleId, categoryId, category, canDelete }) {
+function ArticleCategory({ articleId, categoryId, category, canDelete }) {
   const [anchorEl, setAnchorEl] = useState(null);
   const classes = useStyles();
   const [deleteCategory, { loading: deletingCategory }] = useMutation(
@@ -156,50 +137,9 @@ function Category({ articleId, categoryId, category, canDelete }) {
   );
 }
 
-function ArticleCategories({ articleId, articleCategories }) {
-  const classes = useStyles();
-  const { data: allCategoryData } = useQuery(CATEGORY_LIST_QUERY, {
-    ssr: false,
-  });
-
-  const isInArticle = (articleCategories || []).reduce(
-    (map, { categoryId }) => {
-      map[categoryId] = true;
-      return map;
-    },
-    {}
-  );
-  const otherCategories = (allCategoryData?.ListCategories?.edges || [])
-    .filter(({ node }) => !isInArticle[node.id])
-    .map(({ node }) => node);
-
-  return (
-    <aside>
-      {(articleCategories || []).map(
-        ({ categoryId, category, canUpdateStatus }) => (
-          <Category
-            key={categoryId}
-            category={category}
-            articleId={articleId}
-            categoryId={categoryId}
-            canDelete={canUpdateStatus}
-          />
-        )
-      )}
-
-      {otherCategories.length > 0 && (
-        <Button className={classes.button}>
-          <AddIcon className={classes.buttonIcon} />
-          {t`Add category`}
-        </Button>
-      )}
-    </aside>
-  );
-}
-
-ArticleCategories.fragments = {
-  ArticleCategoriesData,
-  ArticleCategoriesDataForUser,
+ArticleCategory.fragments = {
+  ArticleCategoryData,
+  ArticleCategoryDataForUser,
 };
 
-export default ArticleCategories;
+export default ArticleCategory;
