+++ components/ArticleCategories/index.js
@@ -1,0 +1,3 @@
+import ArticleCategories from './ArticleCategories'
+
+export default ArticleCategories;
