+++ pages/article/[id].js
@@ -50,7 +50,7 @@ const LOAD_ARTICLE = gql`
         }
       }
       articleCategories {
-        ...ArticleCategoriesData
+        ...ArticleCategoryData
       }
     }
   }
@@ -59,7 +59,7 @@ const LOAD_ARTICLE = gql`
   ${CurrentReplies.fragments.CurrentRepliesData}
   ${NewReplySection.fragments.RelatedArticleData}
   ${ArticleItem.fragments.ArticleItem}
-  ${ArticleCategories.fragments.ArticleCategoriesData}
+  ${ArticleCategories.fragments.ArticleCategoryData}
 `;
 
 const LOAD_ARTICLE_FOR_USER = gql`
@@ -73,13 +73,13 @@ const LOAD_ARTICLE_FOR_USER = gql`
         ...ArticleReplyForUser
       }
       articleCategories {
-        ...ArticleCategoriesDataForUser
+        ...ArticleCategoryDataForUser
       }
     }
   }
   ${ReplyRequestReason.fragments.ReplyRequestInfoForUser}
   ${CurrentReplies.fragments.ArticleReplyForUser}
-  ${ArticleCategories.fragments.ArticleCategoriesDataForUser}
+  ${ArticleCategories.fragments.ArticleCategoryDataForUser}
 `;
 
 function ArticlePage() {
